cleaned working v1

+++ database/migrations/2023_08_23_051827_create_sport_user_table.php
@@ -15,8 +15,9 @@ return new class extends Migration
     {
         Schema::create('sport_user', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('sport_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('sport_id');
+
             $table->timestamps();
         });
     }
